Backend with an example of a client talking to it
Back-end (set up as a server) takes in information (one digit at the moment) and multiply it by 2  Written in JavaScript.
Back-end talks to a file called LSU which houses the data model. Then so we can pull from the model it is saved in a file for now.{note: JavaScript is not multi-threaded so it goes line by line(at least from my research)}. Can not figure out a way to make the events run before the end of the file.
also added the code practices.